LayoutTests/http/tests/eventsource/eventsource-cors-with-credentials-expected.txt

CONSOLE MESSAGE: Cannot use wildcard in Access-Control-Allow-Origin when credentials flag is true.
CONSOLE MESSAGE: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-credentials.php?count=1 due to access control checks.
CONSOLE MESSAGE: Cannot use wildcard in Access-Control-Allow-Origin when credentials flag is true.
CONSOLE MESSAGE: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-credentials.php?count=2 due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-credentials.php?count=3 due to access control checks.
Test that EventSource cross-origin requests with credentials fail until the correct CORS headers are sent. Should print a series of PASS messages followed by DONE.

PASS: got error event and readyState is CLOSED
PASS: got error event and readyState is CLOSED
PASS: got error event and readyState is CLOSED
PASS: got cross-origin message event with data "DATA1" and lastEventId "77"
PASS: got cross-origin message event with data "DATA2"
DONE
